Restructurized GpuPicker
