+++ src/js/shot-generator/GPUPickers/Factories/AbstractGPUPickerFactory.js
