+++ src/js/shot-generator/GPUPickers/PickersObjects/EditorPickableCharacter.js
