+++ src/js/shot-generator/GPUPickers/PickersObjects/Pickable.js
