+++ src/js/shot-generator/GPUPickers/PickersObjects/UniversalPickableObject.js
