+++ src/js/shot-generator/GPUPickers/PickersObjects/XRPickableCharacter.js
