+++ src/js/shot-generator/GPUPickers/PickersObjects/XRPickableObjectContainer.js
